Drop the unconditional int32 cast in ParakeetForTDTLoss
The cast is a leftover from ParakeetForRNNTLoss, where it is load
bearing because torchaudio.functional.rnnt_loss requires int32. tdt_loss
is pure torch and has no such requirement, and because the cast was
unconditional it made the labels= path raise on torch below 2.8 for
every label dtype, not just int32 ones.

Keep the int64 cast at the gather as well, so a direct tdt_loss call
with int32 targets keeps working, and widen the test to cover both entry
points and both integer dtypes.

+++ src/transformers/loss/loss_tdt.py
@@ -80,8 +80,8 @@ def tdt_loss(
     blank_log_probs = token_log_probs[:, :, :, blank_token_id]
 
     if max_u > 1:
-        # torch.gather only accepts an int64 index before torch 2.8, and callers
-        # legitimately pass int32 targets (ParakeetForTDTLoss casts to int32 itself).
+        # torch.gather only accepts an int64 index before torch 2.8, and targets can
+        # legitimately arrive as int32 from a caller of tdt_loss.
         targets_expanded = targets.long().unsqueeze(1).expand(-1, max_t, -1)  # (batch, T, U_labels)
         label_log_probs = torch.gather(
             token_log_probs[:, :, : max_u - 1, :],  # (batch, T, U-1, vocab)
@@ -185,7 +185,7 @@ def ParakeetForTDTLoss(
     return tdt_loss(
         token_logits=token_logits,
         duration_logits=duration_logits,
-        targets=labels.to(device).int(),
+        targets=labels.to(device),
         logit_lengths=logit_lengths.to(device).int(),
         target_lengths=label_lengths.to(device).int(),
         blank_token_id=blank_token_id,
